profile implementation fix

+++ WeatherApp/src/main/java/net/hft/dbproject/weatherapp/persistence/UserBaseService.java
@@ -17,6 +17,6 @@ public interface UserBaseService {
     
     public void saveNewUser(UserBase user);
     public UserBase getUserByName(String username);
-    public void updatePasswortByUserId(int userid, String newPassword);
+    public void updatePasswortByUserId(long userid, String newPassword);
    
 }
